requirements.txt
aiohttp>=3.8.1
pandas>=1.4.0
numpy>=1.22.0
websockets>=10.1
influxdb-client>=1.24.0
mysql-connector-python>=8.0.27
redis>=4.1.0
pydantic>=1.9.0
python-dotenv>=0.19.2
matplotlib>=3.5.1
seaborn>=0.11.2
talib-binary>=0.4.24
pytest>=7.0.0
pytest-asyncio>=0.18.0
tqdm>=4.62.3
